backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
requests
selenium>=4.0.0
webdriver-manager
beautifulsoup4
pillow